backend/requirements.txt
Flask==2.3.0
requests==2.31.0
beautifulsoup4==4.12.0
PyGithub==1.59.0
python-dotenv==1.0.0
flask-cors==4.0.0
